Test input on file field
Running tests that interact with file upload fields seems to be a
constant source of problems, so we add a test to specifically exercise
that capability.

STATUS: test fails when using phantomjs, succeeds with
selenium-server (both tested on OS X). For that reason, this probably
won't make it into upstream.